Uploaded SQL code
This is the SQL exploration done in Microsoft SQL Server 2019. The data was initially present in 4 files, which was combined together in SQL with the help of temp tables and analysed to make sense of the numbers. This data is obtained from the Karnataka Government comprising of Mental Health data across a few districts of Karnataka over a period of 4 years (present as 4 files).